+++ tests/engines/netcdf4_engine/test_netcdf4_converter_engine.py
@@ -203,6 +203,37 @@ class TestConverterSimpleNetCDF(ConvertNetCDFBase):
         tiles = tuple(dim.tile for dim in group_schema["x1"].domain)
         assert tiles == (2,)
 
+    @pytest.mark.parametrize("collect_attrs", (True, False))
+    def test_coords_filter(self, netcdf_file, collect_attrs):
+        """Tests all arrays have coords_filters set to desired value."""
+        filters = tiledb.FilterList([tiledb.ZstdFilter(7)])
+        converter = NetCDF4ConverterEngine.from_file(
+            netcdf_file, collect_attrs=collect_attrs, coords_filters=filters
+        )
+        for array_creator in converter.array_creators():
+            assert array_creator.coords_filters == filters
+
+    @pytest.mark.parametrize("collect_attrs", (True, False))
+    def test_offsets_filter(self, netcdf_file, collect_attrs):
+        """Tests all arrays have coords_filters set to desired value."""
+        filters = tiledb.FilterList([tiledb.ZstdFilter(7)])
+        converter = NetCDF4ConverterEngine.from_file(
+            netcdf_file, collect_attrs=collect_attrs, offsets_filters=filters
+        )
+        for array_creator in converter.array_creators():
+            assert array_creator.offsets_filters == filters
+
+    @pytest.mark.parametrize("collect_attrs", (True, False))
+    def test_attrs_filter(self, netcdf_file, collect_attrs):
+        """Tests all arrays have coords_filters set to desired value."""
+        filters = tiledb.FilterList([tiledb.ZstdFilter(7)])
+        converter = NetCDF4ConverterEngine.from_file(
+            netcdf_file, collect_attrs=collect_attrs, attrs_filters=filters
+        )
+        for array_creator in converter.array_creators():
+            for attr_creator in array_creator:
+                assert attr_creator.filters == filters
+
     def test_rename_array(self, netcdf_file):
         converter = NetCDF4ConverterEngine.from_file(netcdf_file, coords_to_dims=False)
         converter.get_array_creator("array0").name = "A1"